feat: 拆分迁移任务执行器代码 #1987
