+++ src/backend/job/biz-job/src/main/kotlin/com/tencent/bkrepo/job/migrate/executor/FinishExecutor.kt
@@ -1,0 +1,81 @@
+/*
+ * Tencent is pleased to support the open source community by making BK-CI 蓝鲸持续集成平台 available.
+ *
+ * Copyright (C) 2024 THL A29 Limited, a Tencent company.  All rights reserved.
+ *
+ * BK-CI 蓝鲸持续集成平台 is licensed under the MIT license.
+ *
+ * A copy of the MIT License is included in this file.
+ *
+ *
+ * Terms of the MIT License:
+ * ---------------------------------------------------
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
+ * documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
+ * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or substantial portions of
+ * the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT
+ * LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN
+ * NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+ * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+ * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package com.tencent.bkrepo.job.migrate.executor
+
+import com.tencent.bkrepo.common.storage.core.StorageProperties
+import com.tencent.bkrepo.common.storage.core.StorageService
+import com.tencent.bkrepo.job.migrate.config.MigrateRepoStorageProperties
+import com.tencent.bkrepo.job.migrate.dao.MigrateFailedNodeDao
+import com.tencent.bkrepo.job.migrate.dao.MigrateRepoStorageTaskDao
+import com.tencent.bkrepo.job.migrate.pojo.MigrateRepoStorageTask
+import com.tencent.bkrepo.job.migrate.pojo.MigrateRepoStorageTaskState
+import com.tencent.bkrepo.job.migrate.utils.ExecutingTaskRecorder
+import com.tencent.bkrepo.repository.api.FileReferenceClient
+import com.tencent.bkrepo.repository.api.RepositoryClient
+import org.slf4j.LoggerFactory
+import org.springframework.stereotype.Component
+
+@Component
+class FinishExecutor(
+    properties: MigrateRepoStorageProperties,
+    storageProperties: StorageProperties,
+    fileReferenceClient: FileReferenceClient,
+    migrateRepoStorageTaskDao: MigrateRepoStorageTaskDao,
+    migrateFailedNodeDao: MigrateFailedNodeDao,
+    storageService: StorageService,
+    executingTaskRecorder: ExecutingTaskRecorder,
+    private val repositoryClient: RepositoryClient,
+) : BaseTaskExecutor(
+    properties,
+    storageProperties,
+    migrateRepoStorageTaskDao,
+    migrateFailedNodeDao,
+    fileReferenceClient,
+    storageService,
+    executingTaskRecorder,
+) {
+    /**
+     * 迁移任务执行结束后对相关资源进行清理
+     */
+    override fun execute(task: MigrateRepoStorageTask): Boolean {
+        with(task) {
+            if (!updateState(task, MigrateRepoStorageTaskState.FINISHING.name)) {
+                return false
+            }
+            logger.info("migrate finished, task[$task]")
+            repositoryClient.unsetOldStorageCredentialsKey(projectId, repoName)
+            migrateRepoStorageTaskDao.removeById(id!!)
+            logger.info("clean migrate task[${projectId}/${repoName}] success")
+            return true
+        }
+    }
+
+    companion object {
+        private val logger = LoggerFactory.getLogger(FinishExecutor::class.java)
+    }
+}
